EleicaoRepositorio

+++ src/main/java/dataa/eleger/repositorios/EleicaoRepositorio.java
@@ -3,5 +3,5 @@ package dataa.eleger.repositorios;
 import dataa.eleger.entidades.EleicaoEntidade;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface EleicaoRepositorio extends PagingAndSortingRepository<EleicaoEntidade, Integer> {
+public interface EleicaoRepositorio extends PagingAndSortingRepository<EleicaoEntidade, Long> {
 }
